Replace CameraController.cs with CameraController2.cs and Rename
Finally deleted the old fild and replaced it with the new one,
updating its name appropriately.
