+++ Assets/Scripts/Sub/CameraController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c1fb0634f5882d549af65bba8a745806
+guid: 3cba36393668e804991ddf12cc21ff35
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
