--- Assets/Scripts/Sub/CameraController2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3cba36393668e804991ddf12cc21ff35
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
